Feature/16/sp redesign (#18)
* Improved the caching and blocking of IPs.
Seems stable when attacked from multiple servers and under a high load.

* Added additional check to prevent hammering the firewall / abuseIPDB

* Update StatisticsController.cs

* Feature/17/console api (#19)

* Added SP.Api.Service base project.
Changed the block caching from keeping the last 3 entries to keeping all entries for the last NN minutes (configurable)

* Add simple cache for IPData requests.
Add additional check to prevent duplicate blocks.

* Corrected cache function for blocking IP ranges.
Added log message to cache cleaning.

* Corrected various merging issues.

* Removed bootstrap
Introduced Fomatic UI
Updated UI for main page.
Statistics page is TODO.

* Updated all NuGet packages.
Various layout related fixes.
SignalR update

* Display the time in local time.
Consistency updates.

+++ Tools/Cleanup/Firewall.cs
@@ -18,28 +18,35 @@ namespace SP.Core
         {
             INetFwPolicy2 fwPolicy2 = (INetFwPolicy2) Activator.CreateInstance(TypeFwPolicy2);
 
-            List<INetFwRule> list = fwPolicy2.Rules.Cast<INetFwRule>()
-                .Where(r => r.Name.ToLowerInvariant().StartsWith("rdp attack")).ToList();
+            List<INetFwRule> list = fwPolicy2?.Rules.Cast<INetFwRule>()
+                .Where(r => r.Name.ToLowerInvariant().StartsWith("sp service block")).ToList();
+
+            if (list == null)
+            {
+	            Console.WriteLine("Unable to find any rules");
+
+	            return;
+            }
 
             Console.WriteLine($"Found {list.Count} rules that should be deleted.");
 
             using StreamWriter writer = new StreamWriter("remove.ps1");
             foreach (INetFwRule rule in list)
             {
-                Console.WriteLine($"Deleting {rule.Name}");
-                writer.WriteLine($"Remove-NetFirewallRule \"{rule.Name}\"");
+	            Console.WriteLine($"Deleting {rule.Name}");
+	            writer.WriteLine($"Remove-NetFirewallRule \"{rule.Name}\"");
 
-                Task.Factory.StartNew(() =>
-                {
-                    try
-                    {
-                        fwPolicy2.Rules.Remove(rule.Name);
-                    }
-                    catch (Exception e)
-                    {
-                        Console.WriteLine(e.Message);
-                    }
-                });
+	            Task.Factory.StartNew(() =>
+	            {
+		            try
+		            {
+			            fwPolicy2.Rules.Remove(rule.Name);
+		            }
+		            catch (Exception e)
+		            {
+			            Console.WriteLine(e.Message);
+		            }
+	            });
             }
         }
     }
